# api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorBody": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "fieldErrors": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Present only for `validation`; keyed by dotted field path.",
            "title": "Fielderrors",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "description": "A machine-readable cause on some 409s — booking: `on_request`, `too_soon`, `sold_out` (`UnbookableReason`).",
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiErrorBody",
        "type": "object"
      },
      "ApiErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
      "Badge": {
        "description": "Computed per departure by `services/catalog/pricing.badge_for` (04 §3).",
        "enum": [
          "filling-fast",
          "sold-out",
          "guaranteed"
        ],
        "title": "Badge",
        "type": "string"
      },
...
      "DealOut": {
        "description": "A running deal (03 R20): a flat amount off per traveller, until the end of `ends_on`.",
        "properties": {
          "endsAt": {
            "description": "When it stops: midnight IST after `endsOn`",
            "format": "date-time",
            "title": "Endsat",
            "type": "string"
          },
          "endsOn": {
            "description": "The last day of the deal, IST",
            "format": "date",
            "title": "Endson",
            "type": "string"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The owner's label; the site says \"Deal\" when null",
            "title": "Label"
          },
          "offPaise": {
            "description": "Flat amount off per traveller (a quote caps it at the traveller's price)",
            "title": "Offpaise",
            "type": "integer"
          },
          "pricePaise": {
            "description": "The price shown beside the struck-through starting price: starting − off",
            "title": "Pricepaise",
            "type": "integer"
          }
        },
        "required": [
          "label",
          "endsOn",
          "endsAt",
          "offPaise",
          "pricePaise"
        ],
        "title": "DealOut",
        "type": "object"
      },
...
      "DepartureList": {
        "properties": {
          "items": {
            "description": "Upcoming departures, soonest first",
            "items": {
              "$ref": "#/components/schemas/DepartureOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DepartureList",
        "type": "object"
      },
...
      "DestinationCard": {
        "properties": {
          "bestMonths": {
            "description": "1-12, in the destination's display order",
            "items": {
              "type": "integer"
            },
            "title": "Bestmonths",
            "type": "array"
          },
          "coverUrl": {
            "title": "Coverurl",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "packageCount": {
            "description": "Live packages",
            "title": "Packagecount",
            "type": "integer"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "startingPricePaise": {
            "description": "Cheapest shown price among its live packages (a running deal counts)",
            "title": "Startingpricepaise",
            "type": "integer"
          },
          "tagline": {
            "title": "Tagline",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "tagline",
          "coverUrl",
          "packageCount",
          "startingPricePaise",
          "bestMonths"
        ],
        "title": "DestinationCard",
        "type": "object"
      },
      "DestinationDetail": {
        "properties": {
          "bestMonths": {
            "description": "1-12",
            "items": {
              "type": "integer"
            },
            "title": "Bestmonths",
            "type": "array"
          },
          "coverUrl": {
            "title": "Coverurl",
            "type": "string"
          },
          "intro": {
            "description": "Markdown",
            "title": "Intro",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "packages": {
            "description": "Live packages, cheapest first",
            "items": {
              "$ref": "#/components/schemas/PackageCard"
            },
            "title": "Packages",
            "type": "array"
          },
          "region": {
            "title": "Region",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "tagline": {
            "title": "Tagline",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "tagline",
          "intro",
          "coverUrl",
          "region",
          "bestMonths",
          "packages"
        ],
        "title": "DestinationDetail",
        "type": "object"
      },
...
      "DestinationList": {
        "properties": {
          "items": {
            "description": "Only destinations with at least 1 live package",
            "items": {
              "$ref": "#/components/schemas/DestinationCard"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "DestinationList",
        "type": "object"
      },
...
      "HomeData": {
        "properties": {
          "deals": {
            "description": "Running deals, ending soonest first; at most 4",
            "items": {
              "$ref": "#/components/schemas/PackageCard"
            },
            "title": "Deals",
            "type": "array"
          },
          "destinations": {
            "description": "Display order, at most 6",
            "items": {
              "$ref": "#/components/schemas/DestinationCard"
            },
            "title": "Destinations",
            "type": "array"
          },
          "packages": {
            "description": "Featured first, then cheapest; at most 6",
            "items": {
              "$ref": "#/components/schemas/PackageCard"
            },
            "title": "Packages",
            "type": "array"
          },
          "stats": {
            "$ref": "#/components/schemas/HomeStats"
          },
          "testimonials": {
            "description": "By position",
            "items": {
              "$ref": "#/components/schemas/TestimonialOut"
            },
            "title": "Testimonials",
            "type": "array"
          }
        },
        "required": [
          "destinations",
          "packages",
          "deals",
          "testimonials",
          "stats"
        ],
        "title": "HomeData",
        "type": "object"
      },
      "HomeStats": {
        "properties": {
          "departures": {
            "description": "Upcoming departures of live packages with seats left",
            "title": "Departures",
            "type": "integer"
          },
          "destinations": {
            "description": "Destinations with at least 1 live package",
            "title": "Destinations",
            "type": "integer"
          },
          "packages": {
            "description": "Live packages",
            "title": "Packages",
            "type": "integer"
          }
        },
        "required": [
          "destinations",
          "packages",
          "departures"
        ],
        "title": "HomeStats",
        "type": "object"
      },
...
      "ImageOut": {
        "properties": {
          "alt": {
            "title": "Alt",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "url",
          "alt",
          "width",
          "height"
        ],
        "title": "ImageOut",
        "type": "object"
      },
...
      "PackageCard": {
        "properties": {
          "badge": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Badge"
              },
              {
                "type": "null"
              }
            ],
            "description": "From the next upcoming departure with seats; sold-out only when all are full"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverurl"
          },
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "deal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DealOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when no deal is running"
          },
          "destination": {
            "description": "Destination display name",
            "title": "Destination",
            "type": "string"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "title": "Highlights",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nights": {
            "title": "Nights",
            "type": "integer"
          },
          "rating": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RatingOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Published reviews only; null when none (B13)"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "startingPricePaise": {
            "title": "Startingpricepaise",
            "type": "integer"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/Theme"
            },
            "title": "Themes",
            "type": "array"
          }
        },
        "required": [
          "slug",
          "name",
          "destination",
          "nights",
          "days",
          "startingPricePaise",
          "themes",
          "coverUrl",
          "highlights",
          "badge",
          "deal",
          "rating"
        ],
        "title": "PackageCard",
        "type": "object"
      },
...
      "PackageDetail": {
        "properties": {
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImageOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "deal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DealOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when no deal is running"
          },
          "departureCity": {
            "title": "Departurecity",
            "type": "string"
          },
          "departures": {
            "description": "Upcoming only, soonest first",
            "items": {
              "$ref": "#/components/schemas/DepartureOut"
            },
            "title": "Departures",
            "type": "array"
          },
          "destination": {
            "$ref": "#/components/schemas/DestinationRef"
          },
          "exclusions": {
            "items": {
              "type": "string"
            },
            "title": "Exclusions",
            "type": "array"
          },
          "faq": {
            "items": {
              "$ref": "#/components/schemas/FaqItem"
            },
            "title": "Faq",
            "type": "array"
          },
          "highlights": {
            "items": {
              "type": "string"
            },
            "title": "Highlights",
            "type": "array"
          },
          "hotels": {
            "items": {
              "$ref": "#/components/schemas/HotelOut"
            },
            "title": "Hotels",
            "type": "array"
          },
          "images": {
            "description": "Gallery order; the cover is first",
            "items": {
              "$ref": "#/components/schemas/ImageOut"
            },
            "title": "Images",
            "type": "array"
          },
          "inclusions": {
            "items": {
              "type": "string"
            },
            "title": "Inclusions",
            "type": "array"
          },
          "itinerary": {
            "items": {
              "$ref": "#/components/schemas/ItineraryDayOut"
            },
            "title": "Itinerary",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nights": {
            "title": "Nights",
            "type": "integer"
          },
          "rating": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RatingOut"
              },
              {
                "type": "null"
              }
            ],
            "description": "Published reviews only, never testimonials; null when none (B13)"
          },
          "related": {
            "description": "Up to 3: same destination, then shared theme",
            "items": {
              "$ref": "#/components/schemas/PackageCard"
            },
            "title": "Related",
            "type": "array"
          },
          "reviews": {
            "description": "The first page of published reviews, newest first; the rest via `/packages/{slug}/reviews`",
            "items": {
              "$ref": "#/components/schemas/PublicReview"
            },
            "title": "Reviews",
            "type": "array"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "startingPricePaise": {
            "description": "Cheapest upcoming double-sharing price; 0 if none",
            "title": "Startingpricepaise",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/Theme"
            },
            "title": "Themes",
            "type": "array"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "slug",
          "name",
          "summary",
          "destination",
          "themes",
          "nights",
          "days",
          "departureCity",
          "startingPricePaise",
          "deal",
          "highlights",
          "inclusions",
          "exclusions",
          "hotels",
          "faq",
          "itinerary",
          "images",
          "cover",
          "departures",
          "related",
          "rating",
          "reviews",
          "updatedAt"
        ],
        "title": "PackageDetail",
        "type": "object"
      },
...
      "PackageList": {
        "properties": {
          "facets": {
            "$ref": "#/components/schemas/SearchFacets"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PackageCard"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "facets"
        ],
        "title": "PackageList",
        "type": "object"
      },
...
      "PublicReviewPage": {
        "description": "`GET /packages/{slug}/reviews?page=N`: newest first, six a page.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PublicReview"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "totalPages": {
            "minimum": 1.0,
            "title": "Totalpages",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "total",
          "totalPages"
        ],
        "title": "PublicReviewPage",
        "type": "object"
      },
...
      "SearchFacets": {
        "description": "What the filter panel offers — derived from the live catalog, never hard-coded.",
        "properties": {
          "budget": {
            "$ref": "#/components/schemas/RangeFacet",
            "description": "Cheapest / priciest shown price (the deal price while one runs) in rupees, rounded out to 1,000; 0/0 if none"
          },
          "destinations": {
            "description": "value = slug; display order",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            },
            "title": "Destinations",
            "type": "array"
          },
          "months": {
            "description": "value = YYYY-MM; upcoming months with a departure that has seats, soonest first",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            },
            "title": "Months",
            "type": "array"
          },
          "nights": {
            "$ref": "#/components/schemas/RangeFacet",
            "description": "Shortest / longest live package; 0/0 when none"
          },
          "themes": {
            "description": "Every theme in enum order; count may be 0",
            "items": {
              "$ref": "#/components/schemas/FacetOption"
            },
            "title": "Themes",
            "type": "array"
          }
        },
        "required": [
          "destinations",
          "themes",
          "months",
          "nights",
          "budget"
        ],
        "title": "SearchFacets",
        "type": "object"
      },
...
      "SortOrder": {
        "enum": [
          "price-asc",
          "price-desc",
          "duration"
        ],
        "title": "SortOrder",
        "type": "string"
      },
...
      "TestimonialOut": {
        "properties": {
          "city": {
            "title": "City",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "packageName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Packagename"
          },
          "packageSlug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when unlinked or the package is not live",
            "title": "Packageslug"
          },
          "rating": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Rating",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "name",
          "city",
          "text",
          "rating",
          "packageSlug",
          "packageName"
        ],
        "title": "TestimonialOut",
        "type": "object"
      },
      "Theme": {
        "enum": [
          "beach",
          "hills",
          "honeymoon",
          "family",
          "adventure",
          "heritage"
        ],
        "title": "Theme",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Tripsmith API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/destinations": {
      "get": {
        "operationId": "listDestinations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Error envelope (06 C0)"
          }
        },
        "summary": "Get Destinations",
        "tags": [
          "public"
        ]
      }
    },
    "/destinations/{slug}": {
      "get": {
        "operationId": "getDestination",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestinationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Error envelope (06 C0)"
          }
        },
        "summary": "Get Destination Route",
        "tags": [
          "public"
        ]
      }
    },
...
    "/home": {
      "get": {
        "operationId": "getHomeData",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeData"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Error envelope (06 C0)"
          }
        },
        "summary": "Get Home Route",
        "tags": [
          "public"
        ]
      }
    },
...
    "/packages": {
      "get": {
        "description": "R3 search. Every filter lives in `search_packages`; this only adds the cache header.",
        "operationId": "searchPackages",
        "parameters": [
          {
            "description": "Destination slugs (≤ 20); any of",
            "in": "query",
            "name": "destination",
            "required": false,
            "schema": {
              "description": "Destination slugs (≤ 20); any of",
              "items": {
                "type": "string"
              },
              "maxItems": 20,
              "title": "Destination",
              "type": "array"
            }
          },
          {
            "description": "Maximum starting price per person, in rupees",
            "in": "query",
            "name": "maxBudget",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "maximum": 10000000,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Maximum starting price per person, in rupees",
              "title": "Maxbudget"
            }
          },
          {
            "in": "query",
            "name": "nightsMin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 30,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nightsmin"
            }
          },
          {
            "in": "query",
            "name": "nightsMax",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 30,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nightsmax"
            }
          },
          {
            "description": "Any of",
            "in": "query",
            "name": "themes",
            "required": false,
            "schema": {
              "description": "Any of",
              "items": {
                "$ref": "#/components/schemas/Theme"
              },
              "title": "Themes",
              "type": "array"
            }
          },
          {
            "description": "YYYY-MM: packages with a departure that month that still has seats",
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM: packages with a departure that month that still has seats",
              "title": "Month"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "price-asc"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Error envelope (06 C0)"
          }
        },
        "summary": "Get Packages",
        "tags": [
          "public"
        ]
      }
    },
    "/packages/{slug}": {
      "get": {
        "operationId": "getPackage",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PackageDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Error envelope (06 C0)"
          }
        },
        "summary": "Get Package Route",
        "tags": [
          "public"
        ]
      }
    },
    "/packages/{slug}/departures": {
      "get": {
        "operationId": "getDeparturesForMonth",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "YYYY-MM; omitted or blank = all upcoming",
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "YYYY-MM; omitted or blank = all upcoming",
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DepartureList"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Error envelope (06 C0)"
          }
        },
        "summary": "Get Departures Route",
        "tags": [
          "public"
        ]
      }
    },
...
    "/packages/{slug}/reviews": {
      "get": {
        "description": "R21: published reviews only, newest first. Page 1 also rides on `GET /packages/{slug}`.",
        "operationId": "listPackageReviews",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "title": "Slug",
              "type": "string"
            }
          },
          {
            "description": "1-based; six a page",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "description": "1-based; six a page",
              "maximum": 500,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicReviewPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Error envelope (06 C0)"
          }
        },
        "summary": "Get Reviews Route",
        "tags": [
          "public"
        ]
      }
    },
...
  }
}